Share widget entry types with the WooCommerce target
Extracts `StoreInfoEntry` / `StoreInfoData` from `StoreInfoProvider.swift`
and `StoreTrendsEntry` from `StoreTrendsWidget.swift` into their own files,
then adds those files to the WooCommerce target's membership exceptions so
`WooCommerceTests` can reach them through `@testable import WooCommerce`.

Refactors the third `StoreTrendsEntryTests` test to construct `StoreInfoData`
directly instead of going through the extension-only
`StoreInfoProvider.placeholderEntry` helper, keeping the same assertions on
the `presentableMetrics` → `WidgetMetricPresenter` → `WidgetReportsURL` chain.
