+++ WooCommerce/WooCommerceTests/Widgets/StoreTrendsEntryTests.swift
@@ -1,5 +1,6 @@
 import Foundation
 import Testing
+import WooFoundation
 @testable import WooCommerce
 
 struct StoreTrendsEntryTests {
@@ -30,20 +31,40 @@ struct StoreTrendsEntryTests {
         #expect(entry.compactRange == StoreStatsWidgetDateRange.lastWeek.localizedCompactRangeLabel)
     }
 
-    @Test func placeholder_entry_provides_chart_trend_and_range_for_trends_metric() throws {
+    @Test func presentable_metric_carries_chart_trend_and_range_aware_deeplink() throws {
+        // Given - a data entry shaped like what the Trends provider hands the rectangular widget:
+        // a single chart-backed metric with previous-period value and an interval series.
+        let currencySettings = CurrencySettings()
+        let chartSeries = (0..<6).map { index in
+            MetricChartPoint(date: Date(timeIntervalSinceReferenceDate: Double(index * 86_400)),
+                             value: Double(80 + index * 10))
+        }
+        let metric = StoreInfoMetric(
+            type: .revenue,
+            value: .currency(Decimal(123_456), currencySettings),
+            previousValue: .currency(Decimal(100_000), currencySettings),
+            chartSeries: chartSeries
+        )
+        let data = StoreInfoData(
+            range: "Last Week",
+            name: "Test Store",
+            revenue: "$123,456",
+            revenueCompact: "$123k",
+            visitors: "67",
+            orders: "23",
+            conversion: "34%",
+            updatedTime: "10:24 PM",
+            metrics: [metric],
+            dateRange: .lastWeek
+        )
+
         // When
-        let entry = StoreInfoProvider.placeholderEntry(dateRange: .lastWeek, metrics: [.revenue])
+        let presentable = try #require(data.presentableMetrics.first)
 
         // Then
-        guard case .data(let data) = entry else {
-            Issue.record("Expected placeholder entry to provide sample data")
-            return
-        }
-
-        let metric = try #require(data.presentableMetrics.first)
         #expect(data.dateRange == .lastWeek)
-        #expect((metric.chartData?.count ?? 0) > 1)
-        #expect(metric.trend?.direction == .up)
-        #expect(metric.tapURL == WidgetReportsURL.url(for: .revenue, range: .lastWeek))
+        #expect((presentable.chartData?.count ?? 0) > 1)
+        #expect(presentable.trend?.direction == .up)
+        #expect(presentable.tapURL == WidgetReportsURL.url(for: .revenue, range: .lastWeek))
     }
 }
